+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Changed startup banner. ([#101](https://github.com/batch-dart/batch.dart/issues/101))
 - Add `onRecover` feature to `RetryConfiguration`. It's executed when all retry has failed. ([#100](https://github.com/batch-dart/batch.dart/issues/100))
 - Supported to output log errors during parallel processing. ([#99](https://github.com/batch-dart/batch.dart/issues/99))
+- Added `onLoadArgs` in the `BatchApplication` callback. ([#102](https://github.com/batch-dart/batch.dart/issues/102))
 
 ## v0.7.1